+++ RemoteControlContainer.m
@@ -51,7 +51,7 @@
 	return NO;	
 }
 
-- (unsigned int) count {
+- (NSUInteger) count {
 	return remoteControls.count;
 }
